Get the new player window showing the data.

+++ win/cocoa/PlayerSelectionWindowController.h
@@ -27,8 +27,8 @@
 
 @interface PlayerSelectionWindowController : NSWindowController <NSWindowDelegate, NSTableViewDataSource, NSTableViewDelegate> {
 
-	int							cntRoles;
-	int							cntRaces;
+	NSInteger					cntRoles;
+	NSInteger					cntRaces;
 		
 	BOOL					*	raceEnabled;
 	BOOL					*	roleEnabled;
